Create multiple addressing modes
Assembly language programs use various addressing modes to access data and perform operations. The three basic addressing modes are:
Register Addressing: In this mode, a register contains the operand. The register may be the first operand, or it can be used as an index or base register.
Immediate Addressing: The operand is stored as part of the instruction itself. This mode is used mostly for constants.
Memory Addressing: In this mode, the operand is stored in memory. The memory address can be specified directly or indirectly using index or base registers.